Move Entity directive to share folder
