+++ src/platforms/dialogflow/google/GoogleAssistantPlatform.ts
@@ -26,6 +26,7 @@ import { ITransition } from "../../../StateMachine";
 import { VoxaApp } from "../../../VoxaApp";
 import { IVoxaEvent } from "../../../VoxaEvent";
 import { IVoxaReply } from "../../../VoxaReply";
+import { Entity } from "../../share/index";
 import { IVoxaPlatformConfig } from "../../VoxaPlatform";
 import { DialogflowPlatform } from "../DialogflowPlatform";
 import { DialogflowReply } from "../DialogflowReply";
@@ -116,6 +117,7 @@ export class GoogleAssistantPlatform extends DialogflowPlatform {
       UpdatePermission,
       Say,
       SessionEntity,
+      Entity,
     ];
   }
 }
