+++ src/platforms/share/index.ts
@@ -1,0 +1,1 @@
+export { Entity, EntityHelper } from "./Entity";
