+++ test/alexa/directives.spec.ts
@@ -6,11 +6,11 @@ import {
   AlexaEvent,
   AlexaPlatform,
   DisplayTemplate,
-  Entity,
   HomeCard,
   IVoxaIntentEvent,
   VoxaApp,
 } from "../../src/";
+import { Entity } from "../../src/platforms/share";
 import { AlexaRequestBuilder } from "./../tools";
 import { variables } from "./../variables";
 import { views } from "./../views";
